refactor: post list code improvements
#32
